refactor: remove unneeded workflows
This commit removes both the now unnecessary GitHub workflow and BitBucket
pipeline.

Originally, the BitBucket repo had a pipeline to automatically trigger a deploy
of the site in Matrix - so every update in GitHub has to be copied across to
BitBucket, where the deploy would then be triggered.

Now this trigger is handled in GitHub with a webhook - so copying-and-pasting
across to BitBucket is no longer necessary.